python/requirements.txt
cython
u-msgpack-python
pytest
sortedcontainers
pytest-benchmark
pytest-xdist
